feat: define api architecture
