+++ README.md
@@ -1,1 +1,13 @@
 # Blockstore api to store blocks and transaction of ethereum, bitcoin chains
+This project is built using deno v2 and kv store without using any dependencies
+
+# Project setup
+
+```
+deno run --allow-net --unstable-kv main.ts
+```
+
+# API's
+
+- Get /api/store
+- Post /api/store
